feat: correct wire format definitions
