+++ pythnet/pythnet_sdk/Cargo.toml
@@ -18,11 +18,9 @@ byteorder = "1.4.3"
 fast-math = "0.1"
 hex = { version = "0.4.3", features = ["serde"] }
 serde = { version = "1.0.144", features = ["derive"] }
-serde_wormhole = { git = "https://github.com/wormhole-foundation/wormhole", tag = "v2.17.1"}
 sha3 = "0.10.4"
 slow_primes = "0.1.14"
 thiserror = "1.0.40"
-wormhole-sdk = { git = "https://github.com/wormhole-foundation/wormhole", tag = "v2.17.1" }
 
 [dev-dependencies]
 base64 = "0.21.0"
@@ -37,6 +35,3 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [build-dependencies]
 rustc_version = "0.4"
-
-[patch.crates-io]
-serde_wormhole = { git = "https://github.com/wormhole-foundation/wormhole", tag = "v2.17.1" }
